fix restconfig debug & boolean handling
The example restconfig is misleading because it has the JSON string
"True" instead of the boolean value true (without quotes).  When
someone tries setting it to "False" it won't work because that's a
valid JSON string that python evaluates to true in places that say
"if config['debug'] ..."

Also fix a typo where the debug value is initialized from the
'insecure' keyword argument.