Filter out inactive copyrights.
Most inactive copyrights appear to be strings (usually source code)
that loosely resemble a copyright statement, but usually aren't.

I'm not sure why the KB has these "inactive" copyrights at all. My
guess is that it *used* to match the copyright pattern, but those
patterns have changed over time and it no longer matches. So instead
of throwing it away, we just mark it as "inactive".